Move OperatorCore under core backend
