+++ devito/core/__init__.py
@@ -7,4 +7,4 @@ backend as well) are used to run Devito on standard CPU architectures.
 # The following used by backends.backendSelector
 from devito.interfaces import DenseData, TimeData  # noqa
 from devito.pointdata import PointData  # noqa
-from devito.operator import Operator  # noqa
+from devito.core.operator import Operator  # noqa
